@@ -4,7 +4,7 @@ namespace Infor\Hiro;
 
 class DIContainer
 {
-  protected $values = array();
+  protected $values = [];
  
   function __set($id, $value)
   {
